* bash_aliases:
mv and cp are using the -i flag per default to prevent clobbering files.

df and free are using the -h flag per default beacuse most of the time there is
a human who wants to read the output.

myip uses echo to append newline character for nicer output.

* bash_functions:

added bak to create a backup for a given file (cp with .bak extension).

added vimg to view images in IMAGEVIEWER.

renamed pdf to vpdf for consistency with vimg.

* bashrc:

IMAGEVIEWER is now set to either eog or BROWSER.