* bash_functions:
Commented out mount_t7. Switching to use the default mount_thunder7 script as
the whole 'tracking' part is not as useful as I thought it would be.

Added wttr to display weather information for the current location.

* bashrc:

Only add anacondas root environment to the PATH instead of a virtual
environment.